requirements.txt
langchain==0.1.0
langchain-community==0.0.13
ollama==0.1.7
pydantic==2.5.0
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.1.0
rich==13.7.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3